Move the potionable items to the potion module

+++ thonkutil-potions-v0/src/main/java/com/jab125/thonkutil/impl/SkipPotionImpl.java
@@ -21,7 +21,7 @@ public class SkipPotionImpl {
     public static boolean contains(Potion potion, int skip) {
         for (var pair : skipList) {
             if (pair.getLeft().equals(potion)) {
-                if (pair.getRight().intValue() == skip) {
+                if (pair.getRight() == skip) {
                     return true;
                 }
             }
